Move MenuOption to a seperate file and share it for both menus
