+++ GreatConsole/ConsoleMenu.cs
@@ -78,18 +78,4 @@ public class ConsoleMenu
 
         return (selectedIndex, options[selectedIndex].name);
     }
-
-    public class MenuOption
-    {
-        public string name;
-        public AllowedColor standardColor;
-        public AllowedColor selectedColor;
-
-        public MenuOption(object _name, AllowedColor _standardColor = DarkYellow, AllowedColor _selectedColor = White)
-        {
-            name = _name.ToString()!;
-            standardColor = _standardColor;
-            selectedColor = _selectedColor;
-        }
-    }
 }
