+++ GreatConsole/ConsoleMultichoiceMenu.cs
@@ -84,18 +84,4 @@ public class ConsoleMultichoiceMenu
             }
         }
     }
-
-    public class MenuOption
-    {
-        public string name;
-        public AllowedColor standardColor;
-        public AllowedColor selectedColor;
-
-        public MenuOption(object _name, AllowedColor _standardColor = DarkYellow, AllowedColor _selectedColor = White)
-        {
-            name = _name.ToString()!;
-            standardColor = _standardColor;
-            selectedColor = _selectedColor;
-        }
-    }
 }
